🎨 Fix:chatbot
벡터 db의 유연성을 위해 기본 추상구조를 만들고, 크로마db가 상속받게 수정
